Show note about Ipopt timing in RTD Documentation (#1)
In the Read the Docs documentation, the JupyterLab and Python script examples print out timing information with the Ipopt output. In some cases, the time to run is a factor of 10 higher than if the example is run on a local machine. The problem is probably due to the speed of the RTD servers. In any event, it's misleading.

This commit has three modifications:

1. In solver.py, a function `print_rtd_note` is called that prints out a note if the environment variable `READTHEDOCS == "True"`, that appears just below the Ipopt timing data.

2. In conf.py, the `READTHEDOCS` variable is injected into the JupyterLab notebooks to trigger the same behavior in notebooks as in scripts. Also, a configuration values, `on_rtd`, is defined that allows conditional content in sphinx documents using the `inconfig` directive.

3. A (conditional) note about this issue was placed on the main index page in the Examples section.